+++ package.json
@@ -17,7 +17,7 @@
     "lint:fix": "eslint --fix",
     "test": "npm run test:unit && npm run test:typescript",
     "test:unit": "c8 --100 node --test",
-    "test:typescript": "tsd"
+    "test:typescript": "tstyche"
   },
   "dependencies": {
     "fastify-plugin": "^5.0.0"
@@ -32,7 +32,7 @@
     "fastify": "^5.0.0",
     "neostandard": "^0.13.0",
     "toad-scheduler": "^3.0.1",
-    "tsd": "^0.33.0"
+    "tstyche": "^7.0.0"
   },
   "homepage": "http://github.com/fastify/fastify-schedule",
   "funding": [
